requirements.txt
streamlit
langchain
langchain-core
langchain-groq
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
tavily-python